Use new quay repo for Conforma task definition
Similar to the previous commit, let's use the non-deprecated repo.

To confirming that it's same image in both places:

    for r in conforma enterprise-contract; do skopeo inspect docker://quay.io/$r/tekton-task:latest | jq -r .Digest; done

But, actually, since this is pinned to quite an old digest, the
digest is not in the new repo, I updated the digest to the current
latest. Note that the old digest appears to be 6 months old, so this
is a going to be a significant upgrade.

For the record, in case we need to troubleshoot a problem:

    $ skopeo inspect --raw docker://quay.io/enterprise-contract/tekton-task@sha256:88755b87abb38842824ed2489e7bb78d3d6fb9179ee046ee6e2542e50a5e24e0 | jq -r '.annotations["org.opencontainers.image.url"]'
    https://github.com/conforma/cli/tree/b5fd51b24ee7ba2532a6aaa45d098bf44a5a8193/tasks/verify-enterprise-contract/0.1/verify-enterprise-contract.yaml tasks/verify-conforma-konflux-ta/0.1/verify-conforma-konflux-ta.yaml

    $ (cd ../cli && git ll -n2 b5fd51b24ee7ba2532a6aaa45d098bf44a5a8193)
    b5fd51b2 6 months ago Simon Baird Merge pull request #2579 from simonbaird/fresh-konflux-pipelines
    27d06474 6 months ago Simon Baird Remove obsolete tekton pipelines

*** Update after rebasing on PR #3046: After rebasing the task
definition digest was updated in that PR instead of this one. The
above comment still applies, but it applies to the commit over in
PR #3046 instead of this commit. I'll leave it here anyway in case
there is some unexpected impact, and we do need to troubleshoot. ***

Ref: https://issues.redhat.com/browse/EC-1570