Use new quay repo for Conforma policy bundles
Update so we're no longer using the old deprecated quay org. Note
that we did rename these repos when copying them to the new org, so
the name variation is correct.

The digests are actually different because we bundle/push
separately, but you can see they are the same like this:

   for r in conforma/build-task-policy enterprise-contract/ec-build_task-policy conforma/task-policy enterprise-contract/ec-task-policy; do skopeo inspect --raw docker://quay.io/$r:latest | jq .annotations; done

Ref: https://issues.redhat.com/browse/EC-1570